# laravel-location-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Location API",
        "version": "3.0.1"
    },
    "paths": {
        "/api/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "List all locations",
                "operationId": "c22aa1cae7ab2b2ad315141d06fec369",
                "responses": {
                    "200": {
                        "description": "Successful location listing"
                    }
                }
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create a new location",
                "operationId": "055665bfbdc18cab3dc956925a9000b4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    },
                                    "color": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location created successfully"
                    }
                }
            }
        },
        "/api/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get a specific location",
                "operationId": "bcaac2b6f85e83cdc99139be535c66e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location retrieved successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update an existing location",
                "operationId": "3ac5d4120b6fb23f55f179649493a2d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "color": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/locations/route": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get all locations sorted by distance from the first location",
                "operationId": "3b3fd038bab40ddc64980e9888841409",
                "responses": {
                    "200": {
                        "description": "Locations retrieved successfully"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "X-Api-Key": {
                "type": "apiKey",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiKeyAuth": []
        }
    ],
    "tags": [
        {
            "name": "Locations",
            "description": "Locations"
        }
    ]
}